pyproject.toml
[project]
name = "thunderstore"
version = "0.2.0"
description = "thunderstore.io modpage support for Mod Organizer 2"
authors = [{ name = "Zash", email = "zash@fantasymail.de" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.4.20250913"
ruff = "^0.13.0"
poethepoet = "^0.37.0"
mypy = "^1.18.1"
pyright = "^1.1.405"
mobase-stubs = "^2.5.2"
pyqt6 = "^6.7.0"

[tool.poe.tasks]
format-imports = "ruff check --select I . --fix"
format-ruff = "ruff format ."
format.sequence = ["format-imports", "format-ruff"]
lint-ruff = "ruff check --extend-select I ."
lint-ruff-format = "ruff format --check ."
lint-pyright = "pyright ."
lint-mypy = "mypy ."
lint.sequence = ["lint-ruff", "lint-ruff-format", "lint-pyright", "lint-mypy"]
lint.ignore_fail = "return_non_zero"

[tool.pyright]
reportMissingModuleSource = false